+++ week3/src/main/java/com/tdd/concert/api/controller/PointController.java
@@ -1,6 +1,6 @@
-package com.tdd.concert.controller;
+package com.tdd.concert.api.controller;
 
-import com.tdd.concert.dto.response.PointResponseDto;
+import com.tdd.concert.api.controller.dto.response.PointResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
