+++ week3/src/main/java/com/tdd/concert/api/controller/dto/request/PaymentRequestDto.java
@@ -1,4 +1,4 @@
-package com.tdd.concert.dto.request;
+package com.tdd.concert.api.controller.dto.request;
 
 import java.time.LocalDate;
