+++ week3/src/main/java/com/tdd/concert/api/controller/dto/request/TokenRequestDto.java
@@ -1,9 +1,6 @@
-package com.tdd.concert.dto.request;
+package com.tdd.concert.api.controller.dto.request;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
-import lombok.Setter;
 import lombok.ToString;
 
 @ToString
