+++ week3/src/main/java/com/tdd/concert/domain/token/component/TokenManager.java
@@ -1,7 +1,7 @@
 package com.tdd.concert.domain.token.component;
 
-import com.tdd.concert.dto.request.TokenRequestDto;
-import com.tdd.concert.dto.response.TokenResponseDto;
+import com.tdd.concert.api.controller.dto.request.TokenRequestDto;
+import com.tdd.concert.api.controller.dto.response.TokenResponseDto;
 
 public interface TokenManager {
