+++ week3/src/test/java/com/tdd/concert/token/component/TokenManagerTest.java
@@ -1,20 +1,10 @@
 package com.tdd.concert.token.component;
 
-import static org.mockito.Mockito.when;
-
 import com.tdd.concert.domain.token.component.TokenGenerator;
-import com.tdd.concert.domain.token.component.TokenManager;
 import com.tdd.concert.domain.token.component.TokenManagerImpl;
 import com.tdd.concert.domain.token.component.TokenValidator;
-import com.tdd.concert.domain.token.model.Token;
-import com.tdd.concert.domain.token.status.ProgressStatus;
-import com.tdd.concert.domain.user.component.UserManager;
 import com.tdd.concert.domain.user.component.UserManagerImpl;
-import com.tdd.concert.domain.user.model.User;
-import com.tdd.concert.dto.request.TokenRequestDto;
-import com.tdd.concert.dto.response.TokenResponseDto;
 import com.tdd.concert.token.mock.MockTokenCoreRepositoryImpl;
-import com.tdd.concert.token.mock.MockUserManagerImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -22,9 +12,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import static org.junit.jupiter.api.Assertions.*;
-
-import java.time.LocalDateTime;
 
 @ExtendWith(MockitoExtension.class)
 public class TokenManagerTest {
